finishing touches, move fixed ICEs to ui tests
