+++ tests/ui/rfcs/rfc-2632-const-trait-impl/assoc-type-const-bound-usage-1.stderr
@@ -1,11 +1,11 @@
 error[E0277]: the trait bound `Trait::{synthetic#0}: Compat` is not satisfied
-  --> $DIR/assoc-type-const-bound-usage-1.rs:16:44
+  --> $DIR/assoc-type-const-bound-usage-1.rs:15:44
    |
 LL | fn unqualified<T: const Trait>() -> Type<{ T::Assoc::func() }> {
    |                                            ^^^^^^^^ the trait `Compat` is not implemented for `Trait::{synthetic#0}`
    |
 note: required by a bound in `Trait::func`
-  --> $DIR/assoc-type-const-bound-usage-1.rs:8:1
+  --> $DIR/assoc-type-const-bound-usage-1.rs:7:1
    |
 LL | #[const_trait]
    | ^^^^^^^^^^^^^^ required by this bound in `Trait::func`
@@ -14,13 +14,13 @@ LL |     fn func() -> i32;
    |        ---- required by a bound in this associated function
 
 error[E0277]: the trait bound `Trait::{synthetic#0}: Compat` is not satisfied
-  --> $DIR/assoc-type-const-bound-usage-1.rs:20:42
+  --> $DIR/assoc-type-const-bound-usage-1.rs:19:42
    |
 LL | fn qualified<T: const Trait>() -> Type<{ <T as Trait>::Assoc::func() }> {
    |                                          ^^^^^^^^^^^^^^^^^^^ the trait `Compat` is not implemented for `Trait::{synthetic#0}`
    |
 note: required by a bound in `Trait::func`
-  --> $DIR/assoc-type-const-bound-usage-1.rs:8:1
+  --> $DIR/assoc-type-const-bound-usage-1.rs:7:1
    |
 LL | #[const_trait]
    | ^^^^^^^^^^^^^^ required by this bound in `Trait::func`
