+++ tests/ui/rfcs/rfc-2632-const-trait-impl/ice-119717-constant-lifetime.rs
@@ -1,0 +1,15 @@
+#![allow(incomplete_features)]
+#![feature(const_trait_impl, effects, try_trait_v2)]
+
+use std::ops::FromResidual;
+
+impl<T> const FromResidual for T {
+    //~^ ERROR const `impl` for trait `FromResidual` which is not marked with `#[const_trait]`
+    //~| type parameter `T` must be used as the type parameter for some local type
+    fn from_residual(t: T) -> _ {
+        //~^ the placeholder `_` is not allowed
+        t
+    }
+}
+
+fn main() {}
