+++ tests/ui/rfcs/rfc-2632-const-trait-impl/ice-119717-constant-lifetime.stderr
@@ -1,0 +1,33 @@
+error: const `impl` for trait `FromResidual` which is not marked with `#[const_trait]`
+  --> $DIR/ice-119717-constant-lifetime.rs:6:15
+   |
+LL | impl<T> const FromResidual for T {
+   |               ^^^^^^^^^^^^
+   |
+   = note: marking a trait with `#[const_trait]` ensures all default method bodies are `const`
+   = note: adding a non-const method body in the future would be a breaking change
+
+error[E0210]: type parameter `T` must be used as the type parameter for some local type (e.g., `MyStruct<T>`)
+  --> $DIR/ice-119717-constant-lifetime.rs:6:6
+   |
+LL | impl<T> const FromResidual for T {
+   |      ^ type parameter `T` must be used as the type parameter for some local type
+   |
+   = note: implementing a foreign trait is only possible if at least one of the types for which it is implemented is local
+   = note: only traits defined in the current crate can be implemented for a type parameter
+
+error[E0121]: the placeholder `_` is not allowed within types on item signatures for functions
+  --> $DIR/ice-119717-constant-lifetime.rs:9:31
+   |
+LL |     fn from_residual(t: T) -> _ {
+   |                               ^ not allowed in type signatures
+   |
+help: try replacing `_` with the type in the corresponding trait method signature
+   |
+LL |     fn from_residual(t: T) -> T {
+   |                               ~
+
+error: aborting due to 3 previous errors
+
+Some errors have detailed explanations: E0121, E0210.
+For more information about an error, try `rustc --explain E0121`.
